[C++] Further clean up of warnings in samples.

+++ aeron-samples/src/main/cpp/BasicPublisher.cpp
@@ -48,7 +48,7 @@ struct Settings
     std::string channel = samples::configuration::DEFAULT_CHANNEL;
     std::int32_t streamId = samples::configuration::DEFAULT_STREAM_ID;
     long long numberOfMessages = samples::configuration::DEFAULT_NUMBER_OF_MESSAGES;
-    long lingerTimeoutMs = samples::configuration::DEFAULT_LINGER_TIMEOUT_MS;
+    int lingerTimeoutMs = samples::configuration::DEFAULT_LINGER_TIMEOUT_MS;
 };
 
 typedef std::array<std::uint8_t, 256> buffer_t;
